pyproject.toml
[project]
name = "booster-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Harold Ruiter", email = "harold@soaratorium.com" }]
requires-python = ">=3.11"
dependencies = [
    "evdev>=1.9.2",
    "numpy>=2.3.5",
    "onnxruntime>=1.18.0",
    "pybind11>=3.0.1",
    "pybind11-stubgen>=2.5.5",
    "pyyaml>=6.0.3",
    "rerun-sdk>=0.27.0",
    "sshkeyboard>=2.3.1",
    "torch>=2.9.1",
]

[project.scripts]
booster-runner = "booster_runner:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
